minimize the map file length, and divide into components
