+++ src/pages/MapPage.tsx
@@ -1,47 +1,17 @@
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import {
-  MapContainer as LeafletMapContainer,
-  TileLayer,
-  Marker,
-  Popup,
-  useMapEvents,
-} from "react-leaflet";
-import {
-  MapPin,
-  Heart,
-  HeartOff,
-  Plus,
-  User,
-  Navigation2,
-  X,
-  Search,
-  MessageCircle,
-} from "lucide-react";
-import {
-  Button,
-  Caption,
-  Subheadline,
-  Title,
-} from "@telegram-apps/telegram-ui";
+import { MapPin } from "lucide-react";
 import { useGeolocation } from "@/hooks/useGeolocation";
 import { retrieveLaunchParams } from "@telegram-apps/sdk-react";
-import { DatabaseLocationSearch } from "@/components/DatabaseLocationSearch";
-import { LocationSearch } from "@/components/LocationSearch";
 import { LocationDetailModal } from "@/components/LocationDetailModal";
-import L from "leaflet";
+import { MapHeader } from "@/components/Map/MapHeader";
+import { MapView } from "@/components/Map/MapView";
+import { FavoritesView } from "@/components/Map/FavoritesView";
+import { MapControls } from "@/components/Map/MapControls";
+import { MapCrosshair } from "@/components/Map/MapCrosshair";
+import { SearchModal } from "@/components/Map/SearchModal";
+import { AddLocationModal } from "@/components/Map/AddLocationModal";
 import "leaflet/dist/leaflet.css";
-
-// Fix for default markers
-delete (L.Icon.Default.prototype as any)._getIconUrl;
-L.Icon.Default.mergeOptions({
-  iconRetinaUrl:
-    "https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.7.1/images/marker-icon-2x.png",
-  iconUrl:
-    "https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.7.1/images/marker-icon.png",
-  shadowUrl:
-    "https://cdnjs.cloudflare.com/ajax/libs/leaflet/1.7.1/images/marker-shadow.png",
-});
 
 interface Location {
   id: number;
@@ -65,40 +35,6 @@ interface AddLocationData {
   description: string;
   type: "permanent" | "temporary";
   category: "grocery" | "restaurant-bar" | "other";
-}
-
-// Custom marker icons for different categories
-const createCustomIcon = (category: string, isFavorite: boolean = false) => {
-  const color =
-    category === "grocery"
-      ? "#22c55e"
-      : category === "restaurant-bar"
-      ? "#f59e0b"
-      : "#6366f1";
-
-  return L.divIcon({
-    html: `<div style="background-color: ${color}; width: 24px; height: 24px; border-radius: 50%; border: 3px solid white; box-shadow: 0 2px 8px rgba(0,0,0,0.3); display: flex; align-items: center; justify-content: center;">
-      ${
-        isFavorite
-          ? '<span style="color: white; font-size: 12px;">♥</span>'
-          : ""
-      }
-    </div>`,
-    className: "custom-marker",
-    iconSize: [24, 24],
-    iconAnchor: [12, 24],
-  });
-};
-
-// Component to setup map reference
-function MapSetup({ setMapRef }: { setMapRef: (map: any) => void }) {
-  const map = useMapEvents({});
-
-  useEffect(() => {
-    setMapRef(map);
-  }, [map, setMapRef]);
-
-  return null;
 }
 
 type TabType = "explore" | "favorites";
@@ -136,10 +72,6 @@ export function MapPage() {
   const { latitude, longitude } = useGeolocation();
   const launchParams = retrieveLaunchParams();
   const telegramUser = (launchParams?.initDataUnsafe as any)?.user;
-
-  // Debug user data
-  console.log("Launch params:", launchParams);
-  console.log("Telegram user:", telegramUser);
 
   const [dynamicMapCenter, setDynamicMapCenter] = useState({
     lat: latitude || 40.7128,
@@ -194,13 +126,6 @@ export function MapPage() {
   };
 
   const handleAddLocation = async () => {
-    console.log("Add location button clicked!", {
-      telegramUser,
-      name: addLocationData.name,
-      data: addLocationData,
-    });
-
-    // Create a fallback user for development/testing
     const effectiveUser = telegramUser || {
       id: 123456789,
       first_name: "Test",
@@ -208,12 +133,7 @@ export function MapPage() {
       username: "testuser",
     };
 
-    if (!addLocationData.name.trim()) {
-      console.log("Validation failed - no name provided");
-      return;
-    }
-
-    console.log("Using effective user:", effectiveUser);
+    if (!addLocationData.name.trim()) return;
 
     setIsSubmitting(true);
     try {
@@ -245,7 +165,7 @@ export function MapPage() {
           type: "permanent",
           category: "other",
         });
-        loadLocations(); // Refresh locations
+        loadLocations();
       } else {
         throw new Error("Failed to add location");
       }
@@ -270,15 +190,12 @@ export function MapPage() {
       if (isFavorited) {
         const response = await fetch(
           `${BACKEND_URL}/api/users/${telegramUser.id}/favorites/${locationId}`,
-          {
-            method: "DELETE",
-          }
+          { method: "DELETE" }
         );
         if (response.ok) {
           setFavoriteLocations((prev) =>
             prev.filter((fav) => fav.id !== locationId)
           );
-          // You could add a toast notification here: "Removed from favorites"
         }
       } else {
         const response = await fetch(
@@ -290,8 +207,7 @@ export function MapPage() {
           }
         );
         if (response.ok) {
-          loadFavorites(); // Refresh favorites
-          // You could add a toast notification here: "Added to favorites"
+          loadFavorites();
         }
       }
     } catch (error) {
@@ -305,7 +221,6 @@ export function MapPage() {
   };
 
   const handleSearchLocationSelect = (location: Location) => {
-    // Navigate to the location
     setDynamicMapCenter({ lat: location.latitude, lng: location.longitude });
     if (mapRef) {
       mapRef.setView([location.latitude, location.longitude], 16);
@@ -320,7 +235,6 @@ export function MapPage() {
     lng: number,
     name: string
   ) => {
-    // Navigate to the location
     setDynamicMapCenter({ lat, lng });
     if (mapRef) {
       mapRef.setView([lat, lng], 16);
@@ -379,7 +293,6 @@ export function MapPage() {
             }
           }
           
-          /* Ensure proper contrast for Telegram */
           .leaflet-popup-content-wrapper {
             background: var(--tg-theme-bg-color) !important;
             color: var(--tg-theme-text-color) !important;
@@ -391,7 +304,6 @@ export function MapPage() {
             background: var(--tg-theme-bg-color) !important;
           }
           
-          /* Better mobile touch targets */
           .leaflet-control-zoom a {
             width: 36px !important;
             height: 36px !important;
@@ -400,6 +312,7 @@ export function MapPage() {
           }
         `}
       </style>
+
       <div
         style={{
           height: "100vh",
@@ -408,1234 +321,81 @@ export function MapPage() {
           backgroundColor: "var(--tg-color-bg)",
         }}
       >
-        {/* Improved Top Header */}
-        <div
-          style={{
-            position: "sticky",
-            top: 0,
-            zIndex: 1000,
-            background: "var(--tg-theme-bg-color)",
-            borderBottom: "1px solid var(--tg-theme-section-separator-color)",
-            boxShadow: "0 2px 8px rgba(0,0,0,0.08)",
-          }}
-        >
-          {/* Main Navigation Tabs */}
-          <div
-            style={{
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "space-between",
-              padding: "12px 16px 8px 16px",
-            }}
-          >
-            <div
-              style={{
-                display: "flex",
-                background: "var(--tg-theme-section-bg-color)",
-                borderRadius: "12px",
-                padding: "4px",
-                border: "1px solid var(--tg-theme-section-separator-color)",
-              }}
-            >
-              <button
-                onClick={() => setActiveTab("explore")}
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  gap: "8px",
-                  padding: "10px 16px",
-                  borderRadius: "8px",
-                  border: "none",
-                  background:
-                    activeTab === "explore"
-                      ? "var(--tg-theme-button-color)"
-                      : "transparent",
-                  color:
-                    activeTab === "explore"
-                      ? "#FFFFFF"
-                      : "var(--tg-theme-text-color)",
-                  fontWeight: "600",
-                  fontSize: "14px",
-                  cursor: "pointer",
-                  transition: "all 0.2s ease",
-                }}
-              >
-                <MapPin size={16} />
-                Explore
-              </button>
-              <button
-                onClick={() => setActiveTab("favorites")}
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  gap: "8px",
-                  padding: "10px 16px",
-                  borderRadius: "8px",
-                  border: "none",
-                  background:
-                    activeTab === "favorites"
-                      ? "var(--tg-theme-button-color)"
-                      : "transparent",
-                  color:
-                    activeTab === "favorites"
-                      ? "#FFFFFF"
-                      : "var(--tg-theme-text-color)",
-                  fontWeight: "600",
-                  fontSize: "14px",
-                  cursor: "pointer",
-                  transition: "all 0.2s ease",
-                }}
-              >
-                <Heart size={16} />
-                Favorites
-              </button>
-            </div>
+        <MapHeader
+          activeTab={activeTab}
+          setActiveTab={setActiveTab}
+          onProfileClick={() => navigate("/profile")}
+          onSearchClick={() => setShowSearchModal(true)}
+        />
 
-            <button
-              onClick={() => navigate("/profile")}
-              style={{
-                background: "var(--tg-theme-section-bg-color)",
-                border: "1px solid var(--tg-theme-section-separator-color)",
-                borderRadius: "50%",
-                width: 40,
-                height: 40,
-                display: "flex",
-                alignItems: "center",
-                justifyContent: "center",
-                cursor: "pointer",
-                color: "var(--tg-theme-text-color)",
-                transition: "all 0.2s ease",
-              }}
-            >
-              <User size={16} />
-            </button>
-          </div>
-
-          {/* Search Bar for Explore Tab */}
-          {activeTab === "explore" && (
-            <div
-              style={{
-                padding: "0 16px 12px 16px",
-              }}
-            >
-              <button
-                onClick={() => setShowSearchModal(true)}
-                style={{
-                  width: "100%",
-                  display: "flex",
-                  alignItems: "center",
-                  gap: "12px",
-                  padding: "14px 20px",
-                  background: "var(--tg-theme-bg-color)",
-                  border: "2px solid var(--tg-theme-section-separator-color)",
-                  borderRadius: "28px",
-                  color: "var(--tg-theme-hint-color)",
-                  fontSize: "16px",
-                  cursor: "pointer",
-                  textAlign: "left",
-                  boxShadow: "0 2px 8px rgba(0,0,0,0.08)",
-                  transition: "all 0.2s ease",
-                }}
-              >
-                <Search size={20} />
-                <span>Search for places, locations...</span>
-              </button>
-
-              {/* Category Pills */}
-              <div
-                style={{
-                  display: "flex",
-                  gap: "8px",
-                  marginTop: "12px",
-                  overflowX: "auto",
-                  paddingBottom: "4px",
-                }}
-              >
-                <div
-                  style={{
-                    display: "flex",
-                    gap: "8px",
-                    whiteSpace: "nowrap",
-                  }}
-                >
-                  <span
-                    style={{
-                      padding: "6px 12px",
-                      background: "var(--tg-theme-section-bg-color)",
-                      borderRadius: "16px",
-                      fontSize: "14px",
-                      color: "var(--tg-theme-text-color)",
-                      border:
-                        "1px solid var(--tg-theme-section-separator-color)",
-                    }}
-                  >
-                    🍽️ Restaurants
-                  </span>
-                  <span
-                    style={{
-                      padding: "6px 12px",
-                      background: "var(--tg-theme-section-bg-color)",
-                      borderRadius: "16px",
-                      fontSize: "14px",
-                      color: "var(--tg-theme-text-color)",
-                      border:
-                        "1px solid var(--tg-theme-section-separator-color)",
-                    }}
-                  >
-                    🛒 Groceries
-                  </span>
-                  <span
-                    style={{
-                      padding: "6px 12px",
-                      background: "var(--tg-theme-section-bg-color)",
-                      borderRadius: "16px",
-                      fontSize: "14px",
-                      color: "var(--tg-theme-text-color)",
-                      border:
-                        "1px solid var(--tg-theme-section-separator-color)",
-                    }}
-                  >
-                    🏪 Other
-                  </span>
-                </div>
-              </div>
-            </div>
-          )}
-        </div>
-
-        {/* Content Area */}
         <div className="flex-1 relative">
           {activeTab === "explore" ? (
-            // Map View for Explore Tab
             <>
-              <LeafletMapContainer
-                center={[dynamicMapCenter.lat, dynamicMapCenter.lng]}
-                zoom={13}
-                style={{ height: "100%", width: "100%" }}
-                zoomControl={true}
-                scrollWheelZoom={true}
-              >
-                <TileLayer
-                  attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-                  url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-                />
+              <MapView
+                center={dynamicMapCenter}
+                locations={locations}
+                favoriteLocations={favoriteLocations}
+                pendingLocation={pendingLocation}
+                userLocation={
+                  latitude && longitude
+                    ? { lat: latitude, lng: longitude }
+                    : null
+                }
+                setMapRef={setMapRef}
+                onLocationClick={handleLocationClick}
+                onToggleFavorite={toggleFavorite}
+              />
 
-                <MapSetup setMapRef={setMapRef} />
+              <MapCrosshair isVisible={isAddLocationMode} />
 
-                {latitude && longitude && (
-                  <Marker position={[latitude, longitude]}>
-                    <Popup>
-                      <div className="text-center">
-                        <Navigation2 className="w-4 h-4 inline text-blue-600 mr-1" />
-                        <span className="font-medium">Your Location</span>
-                      </div>
-                    </Popup>
-                  </Marker>
-                )}
-
-                {locations.map((location) => {
-                  const isFavorited = favoriteLocations.some(
-                    (fav) => fav.id === location.id
-                  );
-                  return (
-                    <Marker
-                      key={location.id}
-                      position={[location.latitude, location.longitude]}
-                      icon={createCustomIcon(location.category, isFavorited)}
-                    >
-                      <Popup>
-                        <div
-                          style={{
-                            minWidth: "240px",
-                            padding: "8px",
-                            background: "var(--tg-theme-bg-color)",
-                            color: "var(--tg-theme-text-color)",
-                          }}
-                        >
-                          <div
-                            style={{
-                              display: "flex",
-                              alignItems: "flex-start",
-                              justifyContent: "space-between",
-                              gap: "8px",
-                              marginBottom: "8px",
-                            }}
-                          >
-                            <h3
-                              style={{
-                                fontWeight: "600",
-                                fontSize: "16px",
-                                margin: 0,
-                              }}
-                            >
-                              {location.name}
-                            </h3>
-                            <button
-                              onClick={() => toggleFavorite(location.id)}
-                              style={{
-                                background: "none",
-                                border:
-                                  "1px solid var(--tg-theme-section-separator-color)",
-                                borderRadius: "50%",
-                                width: "28px",
-                                height: "28px",
-                                display: "flex",
-                                alignItems: "center",
-                                justifyContent: "center",
-                                cursor: "pointer",
-                                color: isFavorited
-                                  ? "#ef4444"
-                                  : "var(--tg-theme-hint-color)",
-                              }}
-                            >
-                              {isFavorited ? (
-                                <Heart size={14} fill="currentColor" />
-                              ) : (
-                                <HeartOff size={14} />
-                              )}
-                            </button>
-                          </div>
-
-                          {location.description && (
-                            <p
-                              style={{
-                                fontSize: "14px",
-                                color: "var(--tg-theme-hint-color)",
-                                marginBottom: "8px",
-                                lineHeight: "1.3",
-                              }}
-                            >
-                              {location.description}
-                            </p>
-                          )}
-
-                          <div
-                            style={{
-                              display: "flex",
-                              alignItems: "center",
-                              gap: "8px",
-                              marginBottom: "12px",
-                            }}
-                          >
-                            <span
-                              style={{
-                                padding: "4px 8px",
-                                borderRadius: "8px",
-                                fontSize: "12px",
-                                fontWeight: "500",
-                                background:
-                                  location.category === "grocery"
-                                    ? "#dcfce7"
-                                    : location.category === "restaurant-bar"
-                                    ? "#fef3c7"
-                                    : "#e0e7ff",
-                                color:
-                                  location.category === "grocery"
-                                    ? "#166534"
-                                    : location.category === "restaurant-bar"
-                                    ? "#92400e"
-                                    : "#3730a3",
-                              }}
-                            >
-                              {location.category.replace("-", " ")}
-                            </span>
-                            <span
-                              style={{
-                                fontSize: "12px",
-                                color: "var(--tg-theme-hint-color)",
-                              }}
-                            >
-                              {location.type}
-                            </span>
-                          </div>
-
-                          <button
-                            onClick={() => handleLocationClick(location)}
-                            style={{
-                              width: "100%",
-                              padding: "8px 12px",
-                              background: "var(--tg-theme-button-color)",
-                              color: "#FFFFFF",
-                              border: "none",
-                              borderRadius: "8px",
-                              fontSize: "14px",
-                              fontWeight: "500",
-                              cursor: "pointer",
-                              display: "flex",
-                              alignItems: "center",
-                              justifyContent: "center",
-                              gap: "6px",
-                            }}
-                          >
-                            <MessageCircle size={14} />
-                            View Details
-                          </button>
-                        </div>
-                      </Popup>
-                    </Marker>
-                  );
-                })}
-
-                {/* Pending location marker */}
-                {pendingLocation && (
-                  <Marker position={[pendingLocation.lat, pendingLocation.lng]}>
-                    <Popup>
-                      <div className="text-center">
-                        <Plus className="w-4 h-4 inline text-green-600 mr-1" />
-                        <span className="font-medium">New Location</span>
-                      </div>
-                    </Popup>
-                  </Marker>
-                )}
-              </LeafletMapContainer>
-
-              {/* Center Crosshair for Add Location Mode */}
-              {isAddLocationMode && (
-                <div
-                  style={{
-                    position: "absolute",
-                    top: "50%",
-                    left: "50%",
-                    transform: "translate(-50%, -50%)",
-                    zIndex: 1000,
-                    pointerEvents: "none",
-                  }}
-                >
-                  <div
-                    style={{
-                      width: "40px",
-                      height: "40px",
-                      position: "relative",
-                    }}
-                  >
-                    {/* Crosshair */}
-                    <div
-                      style={{
-                        position: "absolute",
-                        top: "50%",
-                        left: "50%",
-                        width: "20px",
-                        height: "2px",
-                        backgroundColor: "var(--tg-theme-button-color)",
-                        transform: "translate(-50%, -50%)",
-                        borderRadius: "1px",
-                      }}
-                    />
-                    <div
-                      style={{
-                        position: "absolute",
-                        top: "50%",
-                        left: "50%",
-                        width: "2px",
-                        height: "20px",
-                        backgroundColor: "var(--tg-theme-button-color)",
-                        transform: "translate(-50%, -50%)",
-                        borderRadius: "1px",
-                      }}
-                    />
-                    {/* Center dot */}
-                    <div
-                      style={{
-                        position: "absolute",
-                        top: "50%",
-                        left: "50%",
-                        width: "6px",
-                        height: "6px",
-                        backgroundColor: "var(--tg-theme-button-color)",
-                        borderRadius: "50%",
-                        transform: "translate(-50%, -50%)",
-                        border: "2px solid white",
-                        boxShadow: "0 2px 8px rgba(0,0,0,0.2)",
-                      }}
-                    />
-                  </div>
-                </div>
-              )}
-
-              {/* Floating Action Buttons */}
-              <div
-                style={{
-                  position: "absolute",
-                  bottom: "20px",
-                  right: "20px",
-                  display: "flex",
-                  flexDirection: "column",
-                  gap: "12px",
-                  zIndex: 1000,
-                }}
-              >
-                {/* Add Location Button */}
-                <button
-                  onClick={
-                    isAddLocationMode
-                      ? handleMapCenterAdd
-                      : handleAddLocationModeToggle
+              <MapControls
+                isAddLocationMode={isAddLocationMode}
+                onAddLocationToggle={handleAddLocationModeToggle}
+                onMapCenterAdd={handleMapCenterAdd}
+                onCurrentLocationClick={() => {
+                  if (latitude && longitude) {
+                    setDynamicMapCenter({ lat: latitude, lng: longitude });
+                    if (mapRef) {
+                      mapRef.setView([latitude, longitude], 16);
+                    }
                   }
-                  style={{
-                    width: "56px",
-                    height: "56px",
-                    borderRadius: "50%",
-                    background: isAddLocationMode
-                      ? "var(--tg-theme-destructive-text-color)"
-                      : "var(--tg-theme-button-color)",
-                    color: "white",
-                    border: "none",
-                    boxShadow: "0 4px 16px rgba(0,0,0,0.2)",
-                    cursor: "pointer",
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "center",
-                    transition: "all 0.2s ease",
-                  }}
-                >
-                  {isAddLocationMode ? (
-                    <MapPin size={24} />
-                  ) : (
-                    <Plus size={24} />
-                  )}
-                </button>
-
-                {/* Current Location Button */}
-                {latitude && longitude && (
-                  <button
-                    onClick={() => {
-                      setDynamicMapCenter({ lat: latitude, lng: longitude });
-                      if (mapRef) {
-                        mapRef.setView([latitude, longitude], 16);
-                      }
-                    }}
-                    style={{
-                      width: "48px",
-                      height: "48px",
-                      borderRadius: "50%",
-                      background: "var(--tg-theme-bg-color)",
-                      color: "var(--tg-theme-text-color)",
-                      border:
-                        "1px solid var(--tg-theme-section-separator-color)",
-                      boxShadow: "0 4px 16px rgba(0,0,0,0.15)",
-                      cursor: "pointer",
-                      display: "flex",
-                      alignItems: "center",
-                      justifyContent: "center",
-                      transition: "all 0.2s ease",
-                    }}
-                  >
-                    <Navigation2 size={20} />
-                  </button>
-                )}
-              </div>
-
-              {/* Add Location Mode Instructions */}
-              {isAddLocationMode && (
-                <div
-                  style={{
-                    position: "absolute",
-                    top: "20px",
-                    left: "20px",
-                    right: "20px",
-                    zIndex: 1000,
-                    pointerEvents: "none",
-                  }}
-                >
-                  <div
-                    style={{
-                      background: "var(--tg-theme-bg-color)",
-                      borderRadius: "16px",
-                      padding: "16px",
-                      border:
-                        "1px solid var(--tg-theme-section-separator-color)",
-                      boxShadow: "0 8px 32px rgba(0,0,0,0.15)",
-                    }}
-                  >
-                    <div
-                      style={{
-                        display: "flex",
-                        alignItems: "center",
-                        gap: "12px",
-                      }}
-                    >
-                      <MapPin
-                        size={20}
-                        style={{ color: "var(--tg-theme-button-color)" }}
-                      />
-                      <div>
-                        <div
-                          style={{
-                            fontWeight: "600",
-                            fontSize: "14px",
-                            color: "var(--tg-theme-text-color)",
-                            marginBottom: "4px",
-                          }}
-                        >
-                          Position the crosshair and tap the pin button
-                        </div>
-                        <div
-                          style={{
-                            fontSize: "12px",
-                            color: "var(--tg-theme-hint-color)",
-                          }}
-                        >
-                          Move the map to place your location precisely
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              )}
+                }}
+                hasCurrentLocation={!!(latitude && longitude)}
+              />
             </>
           ) : (
-            // List View for Favorites Tab
-            <div
-              style={{
-                height: "100%",
-                backgroundColor: "var(--tg-color-bg)",
-                overflow: "auto",
-              }}
-            >
-              {favoriteLocations.length === 0 ? (
-                <div
-                  style={{
-                    height: "100%",
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "center",
-                    padding: "2rem",
-                  }}
-                >
-                  <div style={{ textAlign: "center" }}>
-                    <Heart
-                      size={48}
-                      style={{
-                        color: "var(--tg-color-hint-color)",
-                        margin: "0 auto 1rem auto",
-                        display: "block",
-                      }}
-                    />
-                    <Title level="2" style={{ marginBottom: "0.5rem" }}>
-                      No Favorites Yet
-                    </Title>
-                    <Caption>
-                      Explore locations and add them to your favorites to see
-                      them here.
-                    </Caption>
-                  </div>
-                </div>
-              ) : (
-                <div style={{ padding: "1rem" }}>
-                  <div style={{ marginBottom: "1rem", padding: "0 0.5rem" }}>
-                    <Title level="2" style={{ marginBottom: "0.5rem" }}>
-                      Your Favorites ({favoriteLocations.length})
-                    </Title>
-                    <Caption style={{ color: "var(--tg-theme-hint-color)" }}>
-                      Tap any location to view details, ratings, and comments
-                    </Caption>
-                  </div>
-
-                  <div
-                    style={{
-                      display: "flex",
-                      flexDirection: "column",
-                      gap: "0.75rem",
-                    }}
-                  >
-                    {favoriteLocations.map((location) => {
-                      const getCategoryColor = (category: string) => {
-                        switch (category) {
-                          case "grocery":
-                            return "#22c55e";
-                          case "restaurant-bar":
-                            return "#f59e0b";
-                          default:
-                            return "#6366f1";
-                        }
-                      };
-
-                      const getCategoryIcon = (category: string) => {
-                        switch (category) {
-                          case "grocery":
-                            return "🛒";
-                          case "restaurant-bar":
-                            return "🍽️";
-                          default:
-                            return "🏪";
-                        }
-                      };
-
-                      return (
-                        <div
-                          key={location.id}
-                          style={{
-                            backgroundColor: "var(--tg-color-bg-secondary)",
-                            borderRadius: "12px",
-                            padding: "1rem",
-                            border: "1px solid var(--tg-color-separator)",
-                            cursor: "pointer",
-                          }}
-                          onClick={() => handleLocationClick(location)}
-                        >
-                          <div
-                            style={{
-                              display: "flex",
-                              alignItems: "flex-start",
-                              gap: "0.75rem",
-                            }}
-                          >
-                            <div
-                              style={{
-                                backgroundColor: getCategoryColor(
-                                  location.category
-                                ),
-                                borderRadius: "8px",
-                                padding: "8px",
-                                minWidth: "40px",
-                                height: "40px",
-                                display: "flex",
-                                alignItems: "center",
-                                justifyContent: "center",
-                                fontSize: "18px",
-                              }}
-                            >
-                              {getCategoryIcon(location.category)}
-                            </div>
-
-                            <div style={{ flex: 1 }}>
-                              <div
-                                style={{
-                                  display: "flex",
-                                  alignItems: "flex-start",
-                                  justifyContent: "space-between",
-                                  marginBottom: "0.5rem",
-                                }}
-                              >
-                                <Subheadline style={{ fontWeight: "600" }}>
-                                  {location.name}
-                                </Subheadline>
-
-                                <Button
-                                  onClick={(e) => {
-                                    e.stopPropagation(); //adding to Prevent triggering the location click
-                                    toggleFavorite(location.id);
-                                  }}
-                                  mode="plain"
-                                  size="s"
-                                  style={{ padding: "4px", minWidth: "unset" }}
-                                >
-                                  <Heart
-                                    size={16}
-                                    style={{ color: "#ef4444" }}
-                                    fill="currentColor"
-                                  />
-                                </Button>
-                              </div>
-
-                              {location.description && (
-                                <Caption
-                                  style={{
-                                    marginBottom: "0.5rem",
-                                    display: "block",
-                                  }}
-                                >
-                                  {location.description}
-                                </Caption>
-                              )}
-
-                              <div
-                                style={{
-                                  display: "flex",
-                                  alignItems: "center",
-                                  gap: "0.5rem",
-                                  flexWrap: "wrap",
-                                }}
-                              >
-                                <span
-                                  style={{
-                                    backgroundColor: getCategoryColor(
-                                      location.category
-                                    ),
-                                    color: "white",
-                                    fontSize: "12px",
-                                    padding: "2px 8px",
-                                    borderRadius: "6px",
-                                    fontWeight: "500",
-                                  }}
-                                >
-                                  {location.category.replace("-", " ")}
-                                </span>
-
-                                <span
-                                  style={{
-                                    backgroundColor:
-                                      "var(--tg-color-separator)",
-                                    color: "var(--tg-color-text-secondary)",
-                                    fontSize: "12px",
-                                    padding: "2px 8px",
-                                    borderRadius: "6px",
-                                  }}
-                                >
-                                  {location.type}
-                                </span>
-
-                                <Caption
-                                  style={{
-                                    fontSize: "12px",
-                                    color: "var(--tg-color-hint-color)",
-                                  }}
-                                >
-                                  {location.latitude.toFixed(4)},{" "}
-                                  {location.longitude.toFixed(4)}
-                                </Caption>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-                      );
-                    })}
-                  </div>
-                </div>
-              )}
-            </div>
+            <FavoritesView
+              favoriteLocations={favoriteLocations}
+              onLocationClick={handleLocationClick}
+              onToggleFavorite={toggleFavorite}
+            />
           )}
         </div>
 
-        {/* Add Location Modal */}
-        {showAddLocationModal && (
-          <div
-            style={{
-              position: "absolute",
-              inset: 0,
-              backgroundColor: "rgba(15, 23, 42, 0.95)",
-              display: "flex",
-              alignItems: "flex-end",
-              zIndex: 50,
-            }}
-          >
-            <div
-              style={{
-                width: "100%",
-                backgroundColor: "#334155",
-                borderTopLeftRadius: 20,
-                borderTopRightRadius: 20,
-                maxHeight: "85vh",
-                overflow: "auto",
-                boxShadow: "0 -10px 30px rgba(0, 0, 0, 0.5)",
-                animation: "slideUp 0.4s cubic-bezier(0.25, 0.46, 0.45, 0.94)",
-                transform: "translateY(0)",
-              }}
-            >
-              <div
-                style={{
-                  backgroundColor: "#334155",
-                  borderTopLeftRadius: 20,
-                  borderTopRightRadius: 20,
-                  position: "relative",
-                }}
-              >
-                {/* Modal Handle */}
-                <div
-                  style={{
-                    display: "flex",
-                    justifyContent: "center",
-                    padding: "12px 0 8px 0",
-                  }}
-                >
-                  <div
-                    style={{
-                      width: 36,
-                      height: 4,
-                      backgroundColor: "#64748b",
-                      borderRadius: 2,
-                    }}
-                  />
-                </div>
+        <AddLocationModal
+          isOpen={showAddLocationModal}
+          onClose={() => {
+            setShowAddLocationModal(false);
+            setPendingLocation(null);
+          }}
+          addLocationData={addLocationData}
+          setAddLocationData={setAddLocationData}
+          onSubmit={handleAddLocation}
+          isSubmitting={isSubmitting}
+        />
 
-                <div style={{ padding: "0 24px 24px 24px" }}>
-                  <div
-                    style={{
-                      display: "flex",
-                      alignItems: "center",
-                      justifyContent: "space-between",
-                      marginBottom: 24,
-                    }}
-                  >
-                    <Title level="1" style={{ color: "#f1f5f9" }}>
-                      Add Location
-                    </Title>
-                    <button
-                      onClick={() => {
-                        setShowAddLocationModal(false);
-                        setPendingLocation(null);
-                      }}
-                      style={{
-                        background: "none",
-                        border: "none",
-                        cursor: "pointer",
-                        padding: "8px",
-                        borderRadius: "50%",
-                        display: "flex",
-                        alignItems: "center",
-                        justifyContent: "center",
-                        color: "#cbd5e1",
-                      }}
-                    >
-                      <X size={20} />
-                    </button>
-                  </div>
+        <SearchModal
+          isOpen={showSearchModal}
+          onClose={() => setShowSearchModal(false)}
+          searchTab={searchTab}
+          setSearchTab={setSearchTab}
+          onDatabaseLocationSelect={handleSearchLocationSelect}
+          onGlobalLocationSelect={handleGlobalLocationSelect}
+          currentLocation={
+            latitude && longitude ? { lat: latitude, lng: longitude } : null
+          }
+        />
 
-                  <div
-                    style={{
-                      display: "flex",
-                      flexDirection: "column",
-                      gap: 16,
-                    }}
-                  >
-                    <div>
-                      <Subheadline
-                        style={{ marginBottom: 8, color: "#e2e8f0" }}
-                      >
-                        Location Name *
-                      </Subheadline>
-                      <input
-                        type="text"
-                        value={addLocationData.name}
-                        onChange={(e) =>
-                          setAddLocationData((prev) => ({
-                            ...prev,
-                            name: e.target.value,
-                          }))
-                        }
-                        placeholder="Enter location name"
-                        autoFocus
-                        style={{
-                          width: "100%",
-                          padding: "12px 16px",
-                          border: "1px solid #475569",
-                          borderRadius: "12px",
-                          backgroundColor: "#475569",
-                          color: "#f1f5f9",
-                          fontSize: "16px",
-                          outline: "none",
-                        }}
-                      />
-                    </div>
-
-                    <div>
-                      <Subheadline
-                        style={{ marginBottom: 8, color: "#e2e8f0" }}
-                      >
-                        Description
-                      </Subheadline>
-                      <textarea
-                        value={addLocationData.description}
-                        onChange={(e) =>
-                          setAddLocationData((prev) => ({
-                            ...prev,
-                            description: e.target.value,
-                          }))
-                        }
-                        placeholder="Describe this location"
-                        rows={3}
-                        style={{
-                          width: "100%",
-                          padding: "12px 16px",
-                          border: "1px solid #475569",
-                          borderRadius: "12px",
-                          backgroundColor: "#475569",
-                          color: "#f1f5f9",
-                          fontSize: "16px",
-                          outline: "none",
-                          resize: "none",
-                          fontFamily: "inherit",
-                        }}
-                      />
-                    </div>
-
-                    <div
-                      style={{
-                        display: "grid",
-                        gridTemplateColumns: "1fr 1fr",
-                        gap: 16,
-                      }}
-                    >
-                      <div>
-                        <Subheadline
-                          style={{ marginBottom: 8, color: "#e2e8f0" }}
-                        >
-                          Category
-                        </Subheadline>
-                        <select
-                          value={addLocationData.category}
-                          onChange={(e) =>
-                            setAddLocationData((prev) => ({
-                              ...prev,
-                              category: e.target.value as any,
-                            }))
-                          }
-                          style={{
-                            width: "100%",
-                            padding: "12px 16px",
-                            border: "1px solid #475569",
-                            borderRadius: "12px",
-                            backgroundColor: "#475569",
-                            color: "#f1f5f9",
-                            fontSize: "16px",
-                            outline: "none",
-                          }}
-                        >
-                          <option value="grocery">Grocery</option>
-                          <option value="restaurant-bar">Restaurant/Bar</option>
-                          <option value="other">Other</option>
-                        </select>
-                      </div>
-
-                      <div>
-                        <Subheadline
-                          style={{ marginBottom: 8, color: "#e2e8f0" }}
-                        >
-                          Type
-                        </Subheadline>
-                        <select
-                          value={addLocationData.type}
-                          onChange={(e) =>
-                            setAddLocationData((prev) => ({
-                              ...prev,
-                              type: e.target.value as any,
-                            }))
-                          }
-                          style={{
-                            width: "100%",
-                            padding: "12px 16px",
-                            border: "1px solid #475569",
-                            borderRadius: "12px",
-                            backgroundColor: "#475569",
-                            color: "#f1f5f9",
-                            fontSize: "16px",
-                            outline: "none",
-                          }}
-                        >
-                          <option value="permanent">Permanent</option>
-                          <option value="temporary">Temporary</option>
-                        </select>
-                      </div>
-                    </div>
-
-                    <div
-                      style={{
-                        padding: 16,
-                        backgroundColor: "#475569",
-                        borderRadius: 12,
-                        border: "1px solid #64748b",
-                      }}
-                    >
-                      <Caption
-                        style={{
-                          fontWeight: 600,
-                          marginBottom: 4,
-                          color: "#e2e8f0",
-                        }}
-                      >
-                        Coordinates:
-                      </Caption>
-                      <Caption style={{ color: "#cbd5e1" }}>
-                        {addLocationData.lat.toFixed(6)},{" "}
-                        {addLocationData.lng.toFixed(6)}
-                      </Caption>
-                    </div>
-                  </div>
-
-                  <div style={{ marginTop: 24 }}>
-                    <button
-                      type="button"
-                      onClick={(e) => {
-                        e.preventDefault();
-                        e.stopPropagation();
-                        console.log("Button clicked! Event details:", e);
-                        handleAddLocation();
-                      }}
-                      disabled={!addLocationData.name.trim() || isSubmitting}
-                      style={{
-                        width: "100%",
-                        padding: "16px 24px",
-                        backgroundColor: isSubmitting ? "#64748b" : "#3b82f6",
-                        color: "white",
-                        border: "none",
-                        borderRadius: "12px",
-                        fontSize: "16px",
-                        fontWeight: "600",
-                        cursor: isSubmitting ? "not-allowed" : "pointer",
-                        display: "flex",
-                        alignItems: "center",
-                        justifyContent: "center",
-                        gap: "8px",
-                        opacity: isSubmitting ? 0.6 : 1,
-                        transition: "all 0.2s ease",
-                      }}
-                    >
-                      {isSubmitting ? (
-                        "Adding Location..."
-                      ) : (
-                        <>
-                          <Plus size={16} />
-                          Add Location
-                        </>
-                      )}
-                    </button>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-
-        {/* Search Modal */}
-        {showSearchModal && (
-          <div
-            style={{
-              position: "fixed",
-              inset: 0,
-              backgroundColor: "rgba(0, 0, 0, 0.5)",
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "center",
-              zIndex: 2000,
-              padding: "20px",
-            }}
-            onClick={() => setShowSearchModal(false)}
-          >
-            <div
-              style={{
-                width: "100%",
-                maxWidth: "500px",
-                height: "fit-content",
-                maxHeight: "calc(100vh - 40px)",
-                backgroundColor: "var(--tg-theme-bg-color)",
-                borderRadius: "20px",
-                display: "flex",
-                flexDirection: "column",
-                animation: "slideUp 0.3s ease-out",
-                boxShadow: "0 20px 60px rgba(0,0,0,0.3)",
-              }}
-              onClick={(e) => e.stopPropagation()}
-            >
-              {/* Modal Header */}
-              <div
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "space-between",
-                  padding: "20px 20px 0 20px",
-                  borderBottom:
-                    "1px solid var(--tg-theme-section-separator-color)",
-                }}
-              >
-                <Title level="2">Search Locations</Title>
-                <button
-                  onClick={() => setShowSearchModal(false)}
-                  style={{
-                    background: "none",
-                    border: "none",
-                    cursor: "pointer",
-                    padding: "8px",
-                    borderRadius: "50%",
-                    color: "var(--tg-theme-hint-color)",
-                  }}
-                >
-                  <X size={20} />
-                </button>
-              </div>
-
-              {/* Search Tab Selector */}
-              <div
-                style={{
-                  display: "flex",
-                  margin: "16px",
-                  background: "var(--tg-theme-section-bg-color)",
-                  borderRadius: "12px",
-                  padding: "4px",
-                  border: "1px solid var(--tg-theme-section-separator-color)",
-                }}
-              >
-                <button
-                  onClick={() => setSearchTab("db")}
-                  style={{
-                    flex: 1,
-                    padding: "10px 16px",
-                    borderRadius: "8px",
-                    border: "none",
-                    background:
-                      searchTab === "db"
-                        ? "var(--tg-theme-button-color)"
-                        : "transparent",
-                    color:
-                      searchTab === "db"
-                        ? "#FFFFFF"
-                        : "var(--tg-theme-text-color)",
-                    fontWeight: "600",
-                    fontSize: "14px",
-                    cursor: "pointer",
-                    transition: "all 0.2s ease",
-                  }}
-                >
-                  OFM Locations
-                </button>
-                <button
-                  onClick={() => setSearchTab("global")}
-                  style={{
-                    flex: 1,
-                    padding: "10px 16px",
-                    borderRadius: "8px",
-                    border: "none",
-                    background:
-                      searchTab === "global"
-                        ? "var(--tg-theme-button-color)"
-                        : "transparent",
-                    color:
-                      searchTab === "global"
-                        ? "#FFFFFF"
-                        : "var(--tg-theme-text-color)",
-                    fontWeight: "600",
-                    fontSize: "14px",
-                    cursor: "pointer",
-                    transition: "all 0.2s ease",
-                  }}
-                >
-                  Global Search
-                </button>
-              </div>
-              <div
-                style={{
-                  flex: 1,
-                  padding: "0 16px 20px 16px",
-                  overflow: "auto",
-                  minHeight: "300px",
-                }}
-              >
-                {searchTab === "db" ? (
-                  <DatabaseLocationSearch
-                    onLocationSelect={handleSearchLocationSelect}
-                    placeholder="Search stored locations..."
-                    currentLocation={
-                      latitude && longitude
-                        ? { lat: latitude, lng: longitude }
-                        : null
-                    }
-                    showCurrentLocation={true}
-                  />
-                ) : (
-                  <LocationSearch
-                    onLocationSelect={handleGlobalLocationSelect}
-                    placeholder="Search places worldwide..."
-                    currentLocation={
-                      latitude && longitude
-                        ? { lat: latitude, lng: longitude }
-                        : null
-                    }
-                    showCurrentLocation={true}
-                  />
-                )}
-              </div>
-            </div>
-          </div>
-        )}
         {showLocationDetail && selectedLocation && (
           <LocationDetailModal
             location={selectedLocation}
